Installed Pathogen.
Here, I specifically chose Pathogen versus some of the other "fancier"
alternatives like Vundle, NeoBundle, VAM (vim-addon-manager), etc...

Most of said alternatives don't allow for explicit version pinning
and/or manual management of updates (which is a recipe for disaster).
Although git submodules are a bit rough around the edges, they're
perfectly suited to such a use case and, at least in the meantime, it's
all that's needed.