Added vim-sexp and tpope's more sane mappings.
With vim-sexp, we gain much of the power of Emacs' ParEdit for
manipulating s-expressions that are so common with Lisps. Unfortunately,
its default keybindings are pretty garbage and certainly not Vim-esque.

tpope's "vim-sexp-mappings-for-regular-people" provides a compatible set
of (sane) replacement mappings that take advantage of existing Vim
muscle-memory (eg. motions) and, where a new concept is introduced, it's
done in an intuitive way that comes as no surprise.