Added .js ftplugin options for JSX, JSDoc, & Flow.
I use all of these things, so I went ahead and erred on the side of
having them all enabled versus having to remember to enable them at a
later point in time. If it turns out that there's a big performance
penalty *and* I'm infrequently using any of the technologies above
(respectively), I'll opt-out as needed.